+++ src/Providers/AbstractServiceProvider.php
@@ -16,6 +16,7 @@ use Illuminate\Support\ServiceProvider;
 use Namshi\JOSE\JWS;
 use PHPOpenSourceSaver\JWTAuth\Blacklist;
 use PHPOpenSourceSaver\JWTAuth\Claims\Factory as ClaimFactory;
+use PHPOpenSourceSaver\JWTAuth\Console\JWTGenerateCertCommand;
 use PHPOpenSourceSaver\JWTAuth\Console\JWTGenerateSecretCommand;
 use PHPOpenSourceSaver\JWTAuth\Contracts\Providers\Auth;
 use PHPOpenSourceSaver\JWTAuth\Contracts\Providers\JWT as JWTContract;
@@ -78,9 +79,12 @@ abstract class AbstractServiceProvider extends ServiceProvider
         $this->registerPayloadValidator();
         $this->registerClaimFactory();
         $this->registerPayloadFactory();
-        $this->registerJWTCommand();
+        $this->registerJWTCommands();
 
-        $this->commands('tymon.jwt.secret');
+        $this->commands([
+            'tymon.jwt.secret',
+            'tymon.jwt.cert'
+        ]);
     }
 
     /**
@@ -317,9 +321,10 @@ abstract class AbstractServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function registerJWTCommand()
+    protected function registerJWTCommands()
     {
         $this->app->singleton('tymon.jwt.secret', fn () => new JWTGenerateSecretCommand());
+        $this->app->singleton('tymon.jwt.cert', fn () => new JWTGenerateCertCommand());
     }
 
     /**