+++ CHANGELOG.md
@@ -12,6 +12,8 @@ You can find and compare releases at the GitHub release page.
 - Added exception if secret key or private/public key are missing
 
 ### Fixed
+- Add command for generating certs
+
 
 ## [1.4.1] - 2022-01-24